+++ AddOns/!KalielsTracker/Modules/Options.lua
@@ -477,13 +477,12 @@ local options = {
 							order = 2,
 						},
 						frameScrollbar = {
-							name = "顯示捲動指示軸",
-							desc = "啟用捲動功能時顯示捲動指示軸，使用與邊框相同顏色。",
+							name = "顯示捲軸",
+							desc = "啟用捲動功能時顯示捲軸，使用與邊框相同顏色。",
 							type = "toggle",
 							set = function()
 								db.frameScrollbar = not db.frameScrollbar
 								KTF.Bar:SetShown(db.frameScrollbar)
-								KT:SetSize()
 							end,
 							order = 3,
 						},
@@ -1898,7 +1897,7 @@ function KT:CheckAddOn(addon, version, isUI)
 end
 
 function KT:OpenOptions()
-	if self.optionsFrame and not EditModeManagerFrame:IsEditModeActive() then
+	if not self.InCombatBlocked() and self.optionsFrame and not EditModeManagerFrame:IsEditModeActive() then
 		Settings.OpenToCategory(self.optionsFrame.general.name, self.TITLE)
 	end
 end