+++ AddOns/!KalielsTracker/CHANGELOG.md
@@ -1,4 +1,17 @@
 # Kaliel's Tracker changelog
+### 8.1.0
+- ADDED - replace scroll indicator with interactive scrollbar
+- ADDED - TomTom - waypoint support for Collection sources (Appearances, Decors)
+- ADDED - Collections tracking (Appearance, Decor) - dropdown menu, tooltip
+- ADDED - TomTom - support for other TomTom waypoints
+- ADDED - Threat Quest - show POI button
+- CHANGED (help) - Active Patrons
+- CHANGED - addon support - ElvUI 15.05
+- CHANGED - improve rewards in tooltips
+- CHANGED - Libs
+- FIXED - secret number value taint
+- FIXED - Hide visibility rule is ignored when &quot;Hide empty tracker&quot; is enabled
+- FIXED - high CPU usage in Quest tracker (Blizz bug)
 ### 8.0.0
 - ADDED - support for WoW 12.0.0.65512
 - CHANGED - addon support - PetTracker 12.0.0